+++ packages/terra-form-select/src/ComboboxField.jsx
@@ -55,6 +55,12 @@ const propTypes = {
    */
   isLabelHidden: PropTypes.bool,
   /**
+   * Ensure accessibility on touch devices. Will render the dropdown menu in
+   * normal DOM flow with position absolute. By default, the menu renders in a
+   * portal, which is inaccessible on touch devices.
+   */
+  isTouchAccessible: PropTypes.bool,
+  /**
    * Additional attributes to spread onto the label.
    */
   // eslint-disable-next-line react/forbid-prop-types
@@ -106,6 +112,7 @@ const defaultProps = {
   isIncomplete: false,
   isInline: false,
   isInvalid: false,
+  isTouchAccessible: false,
   isLabelHidden: false,
   labelAttrs: {},
   maxWidth: undefined,
@@ -129,6 +136,7 @@ const ComboboxField = ({
   isInline,
   isInvalid,
   isLabelHidden,
+  isTouchAccessible,
   label,
   labelAttrs,
   maxWidth,
@@ -180,6 +188,7 @@ const ComboboxField = ({
         id={selectId}
         isIncomplete={isIncomplete}
         isInvalid={isInvalid}
+        isTouchAccessible={isTouchAccessible}
         defaultValue={defaultValue}
         onChange={onChange}
         placeholder={placeholder}