+++ packages/terra-form-select/src/SelectField.jsx
@@ -61,6 +61,13 @@ const propTypes = {
    */
   isLabelHidden: PropTypes.bool,
   /**
+   * Ensure accessibility on touch devices. Will render the dropdown menu in
+   * normal DOM flow with position absolute. By default, the menu renders in a
+   * portal, which is inaccessible on touch devices. Only applicable to variants
+   * that include an input (e.g. `combobox`, `multiple`, `search`, and `tag`).
+   */
+  isTouchAccessible: PropTypes.bool,
+  /**
    * Additional attributes to spread onto the label.
    */
   // eslint-disable-next-line react/forbid-prop-types
@@ -128,6 +135,7 @@ const defaultProps = {
   isInline: false,
   isInvalid: false,
   isLabelHidden: false,
+  isTouchAccessible: false,
   labelAttrs: {},
   maxSelectionCount: undefined,
   maxWidth: undefined,
@@ -153,6 +161,7 @@ const SelectField = ({
   isInline,
   isInvalid,
   isLabelHidden,
+  isTouchAccessible,
   label,
   labelAttrs,
   maxSelectionCount,
@@ -223,6 +232,7 @@ const SelectField = ({
         id={selectId}
         isIncomplete={isIncomplete}
         isInvalid={isInvalid}
+        isTouchAccessible={isTouchAccessible}
         defaultValue={defaultValue}
         maxSelectionCount={maxSelectionCount !== undefined && maxSelectionCount < 2 ? undefined : maxSelectionCount}
         onChange={onChange}