+++ packages/terra-form-select/src/TagSelectField.jsx
@@ -57,6 +57,12 @@ const propTypes = {
    */
   isLabelHidden: PropTypes.bool,
   /**
+   * Ensure accessibility on touch devices. Will render the dropdown menu in
+   * normal DOM flow with position absolute. By default, the menu renders in a
+   * portal, which is inaccessible on touch devices.
+   */
+  isTouchAccessible: PropTypes.bool,
+  /**
    * Additional attributes to spread onto the label.
    */
   // eslint-disable-next-line react/forbid-prop-types
@@ -112,6 +118,7 @@ const defaultProps = {
   isInline: false,
   isInvalid: false,
   isLabelHidden: false,
+  isTouchAccessible: false,
   labelAttrs: {},
   maxSelectionCount: undefined,
   maxWidth: undefined,
@@ -135,6 +142,7 @@ const TagSelectField = ({
   isInline,
   isInvalid,
   isLabelHidden,
+  isTouchAccessible,
   label,
   labelAttrs,
   maxSelectionCount,
@@ -203,6 +211,7 @@ const TagSelectField = ({
         id={selectId}
         isIncomplete={isIncomplete}
         isInvalid={isInvalid}
+        isTouchAccessible={isTouchAccessible}
         defaultValue={defaultValue}
         maxSelectionCount={maxSelectionCount !== undefined && maxSelectionCount < 2 ? undefined : maxSelectionCount}
         onChange={onChange}