+++ packages/terra-form-select/CHANGELOG.md
@@ -4,6 +4,8 @@ ChangeLog
 Unreleased
 ----------
 ### Added
+* Added `isTouchAccessible` prop for all variants that have an input (e.g. `combobox`, `multiple`, `search`, and `tag`) for mobile accessibility
+  * When used, the dropdown menu is no longer portaled: it is rendered in the normal DOM flow with `position: absolute`
 * Added `isIncomplete` prop to all versions of `SelectField`.
 
 5.31.0 - (September 26, 2019)